lexeme entry improved
